requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-socketio==5.11.0
loguru==0.7.2
jinja2==3.1.3
pydantic==2.6.1
gunicorn==21.2.0
python-dotenv==1.0.1
psutil==5.9.8
mysql-connector-python
